Only T atomic locks, instead of T*N.

Ran on 3060 on Windows 11 with Docker in WSL2.

v1:
```
Loading extension module wkv5...


CUDA warmup...
B 8 T 4096 C 4096 H 64
WARNING:2023-09-06 20:03:37 377:377 init.cpp:143] function cbapi->getCuptiStatus() failed with error CUPTI_ERROR_NOT_INITIALIZED (15)
WARNING:2023-09-06 20:03:37 377:377 init.cpp:144] CUPTI initialization failed - CUDA profiler activities will be missing
INFO:2023-09-06 20:03:37 377:377 init.cpp:146] If you see CUPTI_ERROR_INSUFFICIENT_PRIVILEGES, refer to https://developer.nvidia.com/nvidia-development-tools-solutions-err-nvgpuctrperm-cupti
STAGE:2023-09-06 20:03:37 377:377 ActivityProfilerController.cpp:312] Completed Stage: Warm Up
STAGE:2023-09-06 20:03:37 377:377 ActivityProfilerController.cpp:318] Completed Stage: Collection
STAGE:2023-09-06 20:03:37 377:377 ActivityProfilerController.cpp:322] Completed Stage: Post Processing
B 8 T 4096 C 4096 H 64
STAGE:2023-09-06 20:03:37 377:377 ActivityProfilerController.cpp:312] Completed Stage: Warm Up
STAGE:2023-09-06 20:03:37 377:377 ActivityProfilerController.cpp:318] Completed Stage: Collection
STAGE:2023-09-06 20:03:37 377:377 ActivityProfilerController.cpp:322] Completed Stage: Post Processing
CUDA forward
 ---------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------
           Name    Self CPU %      Self CPU   CPU total %     CPU total  CPU time avg     Self CUDA   Self CUDA %    CUDA total  CUDA time avg    # of Calls
---------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------
          WKV_5        63.73%      65.000us        98.04%     100.000us     100.000us      74.748ms        97.78%      76.405ms      76.405ms             1
    aten::fill_         7.84%       8.000us         7.84%       8.000us       8.000us       1.647ms         2.15%       1.647ms       1.647ms             1
       aten::to         1.96%       2.000us         1.96%       2.000us       0.400us      42.000us         0.05%      42.000us       8.400us             5
    aten::zeros        17.65%      18.000us        34.31%      35.000us      35.000us       6.000us         0.01%       1.657ms       1.657ms             1
    aten::zero_         5.88%       6.000us        13.73%      14.000us      14.000us       3.000us         0.00%       1.650ms       1.650ms             1
---------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------
Self CPU time total: 102.000us
Self CUDA time total: 76.447ms
```

v2:
```
Loading extension module wkv5...


CUDA warmup...
B 8 T 4096 C 4096 H 64
WARNING:2023-09-06 20:06:31 1101:1101 init.cpp:143] function cbapi->getCuptiStatus() failed with error CUPTI_ERROR_NOT_INITIALIZED (15)
WARNING:2023-09-06 20:06:31 1101:1101 init.cpp:144] CUPTI initialization failed - CUDA profiler activities will be missing
INFO:2023-09-06 20:06:31 1101:1101 init.cpp:146] If you see CUPTI_ERROR_INSUFFICIENT_PRIVILEGES, refer to https://developer.nvidia.com/nvidia-development-tools-solutions-err-nvgpuctrperm-cupti
STAGE:2023-09-06 20:06:31 1101:1101 ActivityProfilerController.cpp:312] Completed Stage: Warm Up
STAGE:2023-09-06 20:06:31 1101:1101 ActivityProfilerController.cpp:318] Completed Stage: Collection
STAGE:2023-09-06 20:06:31 1101:1101 ActivityProfilerController.cpp:322] Completed Stage: Post Processing
B 8 T 4096 C 4096 H 64
STAGE:2023-09-06 20:06:31 1101:1101 ActivityProfilerController.cpp:312] Completed Stage: Warm Up
STAGE:2023-09-06 20:06:31 1101:1101 ActivityProfilerController.cpp:318] Completed Stage: Collection
STAGE:2023-09-06 20:06:31 1101:1101 ActivityProfilerController.cpp:322] Completed Stage: Post Processing
CUDA forward
 ---------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------
           Name    Self CPU %      Self CPU   CPU total %     CPU total  CPU time avg     Self CUDA   Self CUDA %    CUDA total  CUDA time avg    # of Calls
---------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------
          WKV_5        71.74%      99.000us        98.55%     136.000us     136.000us      50.551ms        96.89%      52.144ms      52.144ms             1
    aten::fill_         8.70%      12.000us         8.70%      12.000us      12.000us       1.570ms         3.01%       1.570ms       1.570ms             1
       aten::to         1.45%       2.000us         1.45%       2.000us       0.400us      31.000us         0.06%      31.000us       6.200us             5
    aten::zero_         4.35%       6.000us        13.04%      18.000us      18.000us      16.000us         0.03%       1.586ms       1.586ms             1
    aten::zeros        10.87%      15.000us        26.81%      37.000us      37.000us       4.000us         0.01%       1.593ms       1.593ms             1
---------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------
Self CPU time total: 138.000us
Self CUDA time total: 52.175ms
```