@mrsteyk 2023-09-06
And yes, kernel passes the `corectness` test:
```
--> correct = True True , err ratio = 7.253160120711603e-08 7.472804456014708e-08
```